+++ Runtime/DB/DebugUI/DBDebugBlock.cs
@@ -1,0 +1,10 @@
+﻿using TMPro;
+using UnityEngine;
+
+namespace OcDialogue.DB
+{
+    public abstract class DBDebugBlock : MonoBehaviour
+    {
+        public TextMeshProUGUI nameText;
+    }
+}